requirements.txt
flask
python-dotenv
pypdf
sentence-transformers

langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-text-splitters==0.2.4

pinecone==5.0.1
langchain-pinecone==0.1.3

-e .
